Rename components, and update routes for better page hierarchy logic
