+++ src/components/Layouts/DetailsLayout.jsx
@@ -1,4 +1,4 @@
-import { Outlet, NavLink } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import { getSelectedLocation } from '../../redux/search/searchSlice';
 
@@ -12,14 +12,6 @@ const DetailsLayout = () => {
           {' '}
           {name}
         </h2>
-        <nav className="detailsLinks">
-          <NavLink to="all-details">All Details</NavLink>
-          <NavLink to="table">Table</NavLink>
-          <NavLink to="all-charts">All Charts</NavLink>
-          <NavLink to="temperatures">Temperatures</NavLink>
-          <NavLink to="precipitation">Precipitation</NavLink>
-          <NavLink to="humidity">Humidity</NavLink>
-        </nav>
       </section>
 
       <Outlet />
